Reworked to fit into smaller grid

+++ zephyrapp/game/data/prop_assets.c
@@ -4,7 +4,7 @@
 
 const CAT_prop null_prop = {
 	.sprite = &null_sprite,
-	.blockers = (int16_t*[]) {
+	.blockers = (int8_t*[]) {
 	},
 	.blocker_count = 0,
 	.triggers = (struct trigger[]) {
@@ -13,8 +13,8 @@ const CAT_prop null_prop = {
 };
 const CAT_prop reed_prop = {
 	.sprite = &world_reed_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,1,2,2,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,1,2,2,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -29,8 +29,8 @@ const CAT_prop reed_prop = {
 };
 const CAT_prop reed_house_prop = {
 	.sprite = &world_reed_house_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,3,4,5,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,3,4,5,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -45,8 +45,8 @@ const CAT_prop reed_house_prop = {
 };
 const CAT_prop statue_prop = {
 	.sprite = &world_statue_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {1,2,2,3,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {1,2,2,3,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -61,8 +61,8 @@ const CAT_prop statue_prop = {
 };
 const CAT_prop bldg_sm_prop = {
 	.sprite = &world_bldg_sm_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,3,2,5,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,3,2,5,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -71,8 +71,8 @@ const CAT_prop bldg_sm_prop = {
 };
 const CAT_prop bldg_md_prop = {
 	.sprite = &world_bldg_md_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,3,3,5,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,3,3,5,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -81,8 +81,8 @@ const CAT_prop bldg_md_prop = {
 };
 const CAT_prop bldg_lg_prop = {
 	.sprite = &world_bldg_lg_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,3,3,5,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,3,3,5,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -91,8 +91,8 @@ const CAT_prop bldg_lg_prop = {
 };
 const CAT_prop arcade_prop = {
 	.sprite = &world_arcade_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,3,6,5,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,3,6,5,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -107,8 +107,8 @@ const CAT_prop arcade_prop = {
 };
 const CAT_prop market_prop = {
 	.sprite = &world_market_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,2,3,4,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,2,3,4,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -123,8 +123,8 @@ const CAT_prop market_prop = {
 };
 const CAT_prop depot_prop = {
 	.sprite = &world_depot_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,2,3,4,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,2,3,4,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -133,8 +133,8 @@ const CAT_prop depot_prop = {
 };
 const CAT_prop dumpster_prop = {
 	.sprite = &world_dumpster_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,1,1,2,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,1,1,2,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -143,8 +143,8 @@ const CAT_prop dumpster_prop = {
 };
 const CAT_prop trash_prop = {
 	.sprite = &world_trash_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,0,1,1,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,0,1,1,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -153,8 +153,8 @@ const CAT_prop trash_prop = {
 };
 const CAT_prop bench_prop = {
 	.sprite = &world_bench_tile_sprite,
-	.blockers = (int16_t*[]) {
-		(int16_t[]) {0,0,1,1,},
+	.blockers = (int8_t*[]) {
+		(int8_t[]) {0,0,1,1,},
 	},
 	.blocker_count = 1,
 	.triggers = (struct trigger[]) {
@@ -163,7 +163,7 @@ const CAT_prop bench_prop = {
 };
 const CAT_prop track_prop = {
 	.sprite = &floor_grass_tile_sprite,
-	.blockers = (int16_t*[]) {
+	.blockers = (int8_t*[]) {
 	},
 	.blocker_count = 0,
 	.triggers = (struct trigger[]) {
